+++ crankshaft-engine/CHANGELOG.md
@@ -10,3 +10,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 * Adds the initial version of the crate.
+* Adds basic auth to the TES examples
+  ([[#6](https://github.com/stjude-rust-labs/crankshaft/issues/6)]).